Add new lines, and refactor.
- Extracted function declarations into krida_i2c_dimmer.cpp
- Renamed namespaces to match esphome style guide
- Added new lines to end of files
