+++ esphome/components/krida_4ch_dimmer/output.py
